(feat): refactor names entities
change for semantics names TopicRouter TopicModel because init one new fork
